+++ portal/tests/test_teacher_student.py
@@ -579,7 +579,6 @@ class TestTeacherStudentFrontend(BaseTest):
         create_organisation_directly(email)
         _, _, access_code = create_class_directly(email)
         student_name_1, _, student = create_school_student_directly(access_code)
-        _, _, _ = create_school_student_directly(access_code)
 
         self.selenium.get(self.live_server_url)
         page = (
@@ -608,3 +607,44 @@ class TestTeacherStudentFrontend(BaseTest):
         logs = JoinReleaseStudent.objects.filter(student=student)
         assert len(logs) == 1
         assert logs[0].action_type == JoinReleaseStudent.RELEASE
+
+    def test_multiple_dismiss(self):
+        email, password = signup_teacher_directly()
+        create_organisation_directly(email)
+        _, _, access_code = create_class_directly(email)
+        student_name_1, _, student = create_school_student_directly(access_code)
+        student_name_2, _, student_2 = create_school_student_directly(access_code)
+
+        self.selenium.get(self.live_server_url)
+        page = (
+            HomePage(self.selenium)
+            .go_to_teacher_login_page()
+            .login(email, password)
+            .open_classes_tab()
+            .go_to_class_page()
+        )
+        assert page.student_exists(student_name_1)
+        assert page.student_exists(student_name_2)
+
+        page = page.toggle_all_students()
+        page = page.dismiss_students()
+
+        # dismiss with the same email address
+        SAME_EMAIL = "student_email@gmail.com"
+        page = page.enter_email(SAME_EMAIL, 0)
+        page = page.enter_email(SAME_EMAIL, 1)
+        page = page.dismiss()
+
+        # the first should be released, the second not
+        assert not page.student_exists(student_name_1)
+        assert page.student_exists(student_name_2)
+
+        # dismiss using teacher's email
+        page = page.toggle_all_students()
+        page = page.dismiss_students()
+
+        page = page.enter_email(email, 0)
+        page = page.dismiss()
+
+        # student should still exist
+        assert page.student_exists(student_name_2)